requirements.txt
flask~=2.3.3
flasgger~=0.9.7.1
zeep~=4.2.1
requests~=2.28.1
onvif-zeep
time
pillow
socket
random
string